ov006: promote dScMgD3DBase_c into a single C++ TU
Gathers 27 one-function shards -- the 3D minigame base class's own
seventeen members plus the ten free helpers that shared its translation
unit -- into src/actors/dScMgD3DBase_c.cpp.  No header change was needed:
include/dScMgD3DBase_c.h already declared every override and already had
the destructor inline and declared first.

  tubuild verify   : 27/27 MATCH, objisolate clean, reloc-destinations
                     clean, emission order ROM-ascending
  tubuild linkcheck: SCRATCH-LINK-VERIFIED (pre-promotion run), the whole
                     0x020e6c28..0x020e7660 range reproducing from one
                     object, 106/106 modules exact, ROM sha256 identical
                     to the pristine-tree baseline control
  romdata_check    : 11 symbols, 6 VERIFIED, 5 PARTIAL, 0 DIFFERS

ELEVEN compiler_only_output rows, DERIVED rather than copied from the
queue's estimate: dScMgD3DBase_c -> dScMgBase_c -> dScene_c -> dBase_c ->
fBase_c is five levels, so 2 x 5 typeinfo rows + 1 vtable = 11, with no
Vector3 D1 term because the class has no Vector3 member.  Each row's
canonical module was read out of the module's own symbols.txt rather than
assumed to be ov006 -- eight of the eleven live in arm9 or ov004.

The two forcing-destructor scaffolds the D0/D1 shards carried are deleted
rather than ported.  This TU defines the key function
(BeforeInitResources, the first declared non-inline virtual), so mwcc
emits the vtable here and slots 16 and 17 drag the D1/D0 pair out of line
in that order -- which is the retail order.

Twelve declaration conflicts were reconciled, five of which tubuild's
detector cannot see because it compares declaration text and not
parameter types.  One was settled against the ROM: retail
AfterCleanupResources has `mov r0, r5` before `bl 0x020e7110`, so that
call really does take an argument its callee ignores.  All twelve are
written up at the head of the new file and recorded in the manifest.
