+++ config/arm9/overlays/ov006/delinks.txt
@@ -2507,113 +2507,9 @@ src/d_s_mg_curling2.c:
     complete
     .text start:0x020e6bf4 end:0x020e6c28
 
-src/_ZN14dScMgD3DBase_cD1Ev.cpp:
+src/actors/dScMgD3DBase_c.cpp:
     complete
-    .text start:0x020e6c28 end:0x020e6c60
-
-src/_ZN14dScMgD3DBase_cD0Ev.cpp:
-    complete
-    .text start:0x020e6c60 end:0x020e6cac
-
-src/_ZN14dScMgD3DBase_c25OnAimedAtWithEggReturnVecEv.cpp:
-    complete
-    .text start:0x020e6cac end:0x020e6d24
-
-src/_ZN14dScMgD3DBase_c16OnAimedAtWithEggEv.cpp:
-    complete
-    .text start:0x020e6d24 end:0x020e6d8c
-
-src/_ZN14dScMgD3DBase_c19OnHitFromUnderneathEv.cpp:
-    complete
-    .text start:0x020e6d8c end:0x020e6d98
-
-src/_ZN14dScMgD3DBase_c15OnHitByMegaCharEv.cpp:
-    complete
-    .text start:0x020e6d98 end:0x020e6da4
-
-src/func_ov006_020e6da4.c:
-    complete
-    .text start:0x020e6da4 end:0x020e6db4
-
-src/func_ov006_020e6db4.c:
-    complete
-    .text start:0x020e6db4 end:0x020e6df0
-
-src/Sound_PlayBank1Panned.cpp:
-    complete
-    .text start:0x020e6df0 end:0x020e6e3c
-
-src/func_ov006_020e6e3c.c:
-    complete
-    .text start:0x020e6e3c end:0x020e6e4c
-
-src/_ZN14dScMgD3DBase_c24OnHitByCannonBlastedCharEv.cpp:
-    complete
-    .text start:0x020e6e4c end:0x020e6e54
-
-src/_ZN14dScMgD3DBase_c8OnPushedEv.cpp:
-    complete
-    .text start:0x020e6e54 end:0x020e6e78
-
-src/_ZN14dScMgD3DBase_c8OnKickedEv.cpp:
-    complete
-    .text start:0x020e6e78 end:0x020e6f60
-
-src/_ZN14dScMgD3DBase_c21AfterCleanupResourcesEj.cpp:
-    complete
-    .text start:0x020e6f60 end:0x020e700c
-
-src/_ZN14dScMgD3DBase_c11AfterRenderEj.cpp:
-    complete
-    .text start:0x020e700c end:0x020e7040
-
-src/_ZN14dScMgD3DBase_c12BeforeRenderEv.cpp:
-    complete
-    .text start:0x020e7040 end:0x020e7074
-
-src/_ZN14dScMgD3DBase_c14BeforeBehaviorEv.cpp:
-    complete
-    .text start:0x020e7074 end:0x020e70c0
-
-src/_ZN14dScMgD3DBase_c18AfterInitResourcesEj.cpp:
-    complete
-    .text start:0x020e70c0 end:0x020e70e4
-
-src/_ZN14dScMgD3DBase_c19BeforeInitResourcesEv.cpp:
-    complete
-    .text start:0x020e70e4 end:0x020e7110
-
-src/func_ov006_020e7110.c:
-    complete
-    .text start:0x020e7110 end:0x020e7124
-
-src/_ZN14dScMgD3DBase_c9Virtual84Ev.cpp:
-    complete
-    .text start:0x020e7124 end:0x020e72c0
-
-src/_ZN14dScMgD3DBase_c9Virtual7CEv.cpp:
-    complete
-    .text start:0x020e72c0 end:0x020e73c4
-
-src/func_ov006_020e73c4.cpp:
-    complete
-    .text start:0x020e73c4 end:0x020e740c
-
-src/func_ov006_020e740c.c:
-    complete
-    .text start:0x020e740c end:0x020e7428
-
-src/func_ov006_020e7428.c:
-    complete
-    .text start:0x020e7428 end:0x020e7508
-
-src/func_ov006_020e7508.cpp:
-    complete
-    .text start:0x020e7508 end:0x020e759c
-
-src/func_ov006_020e759c.c:
-    complete
-    .text start:0x020e759c end:0x020e7660
+    .text start:0x020e6c28 end:0x020e7660
 
 src/_ZN12dScMg3DEsp_cD1Ev.cpp:
     complete
