+++ config/converted-baseline.json
@@ -191,7 +191,6 @@
     "src/ShowArea.c",
     "src/Snd_SendCommand.c",
     "src/SoftReset.c",
-    "src/Sound_PlayBank1Panned.cpp",
     "src/Sound_PlayIfNotActive.c",
     "src/Sqrt.c",
     "src/StartExitCharacterWipe.c",
@@ -787,9 +786,6 @@
     "src/_ZN14dScMgCurling_c8BehaviorEv.cpp",
     "src/_ZN14dScMgCurling_cD0Ev.cpp",
     "src/_ZN14dScMgCurling_cD1Ev.cpp",
-    "src/_ZN14dScMgD3DBase_c24OnHitByCannonBlastedCharEv.cpp",
-    "src/_ZN14dScMgD3DBase_cD0Ev.cpp",
-    "src/_ZN14dScMgD3DBase_cD1Ev.cpp",
     "src/_ZN14dScMgMemory2_c15OnGroundPoundedEv.cpp",
     "src/_ZN14dScMgMemory2_c6RenderEv.cpp",
     "src/_ZN14dScMgMemory2_cD0Ev.cpp",
@@ -2185,6 +2181,10 @@
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cC2Ev",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cD0Ev",
     "src/actors/dBgActor_c.cpp#_ZN10dBgActor_cD1Ev",
+    "src/actors/dScMgD3DBase_c.cpp#Sound_PlayBank1Panned",
+    "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_c24OnHitByCannonBlastedCharEv",
+    "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_cD0Ev",
+    "src/actors/dScMgD3DBase_c.cpp#_ZN14dScMgD3DBase_cD1Ev",
     "src/engine/fader/_ZN10FaderColorD0Ev.cpp",
     "src/engine/fader/_ZN10FaderColorD1Ev.cpp",
     "src/engine/fader/_ZN15FaderBrightness10SetToStartEv.cpp",
