--- src/_ZN14dScMgD3DBase_c19BeforeInitResourcesEv.cpp
@@ -1,13 +1,0 @@
-//cpp
-// @symbol _ZN14dScMgD3DBase_c19BeforeInitResourcesEv
-/* recovered: named members + shared header, real C++ method, declarations from a shared header */
-#include "decl_common.h"
-/* recovered: named members + shared header, real C++ method */
-#include "dScMgD3DBase_c.h"
-
-bool dScMgD3DBase_c::BeforeInitResources()
-{
-  if(_ZN11dScMgBase_c19BeforeInitResourcesEv(((void*)this))==0) return 0;
-  unk_5000 = 0;
-  return 1;
-}
