--- src/func_ov006_020e6da4.c
@@ -1,8 +1,0 @@
-/* func_ov006_020e6da4 at 0x020e6da4 - thunk: func_020126e8(a + 0x80000) */
-
-extern int func_020126e8(int a);
-
-int func_ov006_020e6da4(int a)
-{
-    return func_020126e8(a + 0x80000);
-}
