--- src/func_ov006_020e6db4.c
@@ -1,6 +1,0 @@
-void func_ov006_020e6db4(int a0,int a1,int a2){
-  extern int func_020126e8(int a);
-  extern void func_020126ac(int a0,int a1,int a2,int a3,int s0);
-  int s0=func_020126e8(a1+0x80000);
-  func_020126ac(a0,6,0,a2,s0);
-}
