--- src/func_ov006_020e7110.c
@@ -1,2 +1,0 @@
-extern char data_0209e660[];
-void func_ov006_020e7110(void) { data_0209e660[0] = 0; }
