--- src/func_ov006_020e740c.c
@@ -1,7 +1,0 @@
-extern void func_02056674(const void *src, unsigned int offset, unsigned int count);
-extern char data_ov006_02141a54;
-
-void func_ov006_020e740c(void)
-{
-    func_02056674(&data_ov006_02141a54, 0, 0x400u);
-}
